requirements.txt
streamlit==1.28.0
langchain==0.0.350
langchain-community==0.0.3
faiss-cpu==1.7.4
ollama==0.1.7
pandas==2.1.3
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2
python-dotenv==1.0.0
streamlit-chat==0.1.1
numpy==1.24.3
pickle-mixin==1.0.2
